log: make progress with use of stroke
When using the stroke draw option over the fill option
the x, y, width and height values need to be changed
a little bit.

This is due to the way that the stoke is drawn. The
stroke effect grows outward/inward.

For example: if a square is located at (5,5) with
a width and height of (5,5) and has a stroke width
of 2. The square appears as if it's top left corner
was located at (4,4) and with a width and height of
(6,6).

The new implementation is successful for the most part
but it needs some minor work.