log: improve Box class
I believe I've fixed most if not all of the previous
issues with the use of a stroke to draw a box.

Groups now apply the correct rect object to boxes and
they stack nicely.

There is one minor issue but this is more of a QOL thing
than an actual bug. When stroked boxes are placed
in a group the width of their strokes are doubled along
edges. This may not be the intended behavior for some
instances.