log: test array visual
I made a "mock" visualization for an Array. Improvements
need to be made, and the system needs to be developed
further, but it works as intended for now.

Also in this commit, I chose to go with a "border"
type of thing on a Group object in order to overcome
the issue of Boxes using stroke as it would double
the size of an edge. To overcome this I instead
use the Group's gap and border properties to simulate
the appearance of stroked Boxes.